+++ .gitignore
@@ -12,6 +12,9 @@ node_modules
 # production
 dist
 
+# cache
+.cache
+
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
