+++ package.json
@@ -36,7 +36,8 @@
     "@babel/preset-typescript": "7.24.7",
     "@patternfly/react-icons": "5.3.2",
     "@redhat-cloud-services/eslint-config-redhat-cloud-services": "2.0.4",
-    "@redhat-cloud-services/frontend-components-config": "5.0.5",
+    "@redhat-cloud-services/frontend-components-config": "6.0.16",
+    "@redhat-cloud-services/tsc-transform-imports": "1.0.11",
     "@rtk-query/codegen-openapi": "1.2.0",
     "@testing-library/dom": "10.3.2",
     "@testing-library/jest-dom": "6.4.6",
@@ -79,6 +80,7 @@
     "stylelint": "15.11.0",
     "stylelint-config-recommended-scss": "13.1.0",
     "ts-node": "10.9.2",
+    "ts-patch": "3.2.0",
     "typescript": "5.5.3",
     "uuid": "10.0.0",
     "vitest": "1.6.0",
@@ -88,23 +90,20 @@
   },
   "scripts": {
     "lint": "npm-run-all lint:*",
-    "lint:js": "eslint config src",
-    "lint:js:fix": "eslint config src --fix",
+    "lint:js": "eslint src",
+    "lint:js:fix": "eslint src --fix",
     "lint:sass": "stylelint 'src/**/*.scss' --config .stylelintrc.json",
-    "devel": "webpack serve --config config/devel.webpack.config.js",
-    "prod-beta": "BETA=true PROXY=true webpack serve --config config/dev.webpack.config.js",
-    "prod-stable": "PROXY=true webpack serve --config config/dev.webpack.config.js",
-    "stage-stable": "STAGE=true npm run prod-stable",
-    "stage-beta": "STAGE=true npm run prod-beta",
-    "stage-beta:msw": "MSW=TRUE npm run stage-beta",
+    "start": "fec dev",
+    "start:msw": "MSW=TRUE fec dev",
     "test": "TZ=UTC vitest run",
     "test:watch": "TZ=UTC vitest",
     "test:coverage": "TZ=UTC vitest run --coverage",
-    "build": "webpack --config config/prod.webpack.config.js",
+    "build": "fec build",
     "api": "npm-run-all api:pull api:generate",
     "api:generate": "bash api/codegen.sh",
     "api:pull": "bash api/pull.sh",
-    "verify": "npm-run-all build lint test"
+    "verify": "npm-run-all build lint test",
+    "postinstall": "ts-patch install"
   },
   "insights": {
     "appname": "image-builder"
